Demonstrate references tab rendering issue
flake 8
